Go back to the non-ptr BufInfo & AttrInfo
I think you do want local copies of each of these, probably.
The important thing is that the pointer's lifetime is valid
throughout the call to renderer_render...

Also, the calls to SetAttrInfo and SetBufInfo could likely be
optimized to only occur e.g. 'once' per set of meshes that
use the same attr info.